transformers/tests/requirements.txt
aistore[etl]>=1.15.0
filetype
keras
numpy
pillow
pyyaml
requests
scikit-image
scipy
pytest
tensorflow
soundfile
opencv-python
aiofiles
kaggle
typing-extensions>=4.3.0